+++ engine/utils/project_utils.js
@@ -423,4 +423,66 @@ ProjectUtils.installError = function (cli, err) {
     cli.ok('====================================================');
 };
 
+/**
+ * Unpublishes a project from the repository.
+ *
+ * @param {Project} project The project to be unpublished.
+ * @param {Object} promptResult The password and other questions asked by the application.
+ * @return Promise
+ */
+ProjectUtils.unpublishProject = function (project, promptResult) {
+    var deferred = Q.defer(),
+        data = {
+            name: project.get('name'),
+            user: project.get('author').email,
+            pass: promptResult.Password
+        };
+
+    /* The check of the dependencies is being done on the server. */
+    unirest.delete(OmenAPI.buildURL('/unpublish/project'))
+        .type('json')
+        .send(data)
+        .end(function (response) {
+            if (response.statusType == 4 || response.statusType == 5)
+                deferred.reject(new Error({status: response.status, body: response.body}));
+            else {
+                deferred.resolve(response.body);
+            }
+        });
+
+    return deferred.promise;
+};
+
+/**
+ * Performs a check of the dependencies on the server.
+ *
+ * @param {Project} project The project to be unpublished.
+ * @param {String} version The version to be unpublished.
+ * @param {Object} promptResult The password and other questions asked by the application.
+ * @return Promise
+ */
+ProjectUtils.unpublishVersion = function (project, version, promptResult) {
+    var deferred = Q.defer(),
+        data = {
+            name: project.get('name'),
+            version: version,
+            user: project.get('author').email,
+            pass: promptResult.Password
+        };
+
+    /* The check of the dependencies is being done on the server. */
+    unirest.delete(OmenAPI.buildURL('/unpublish/version'))
+        .type('json')
+        .send(data)
+        .end(function (response) {
+            if (response.statusType == 4 || response.statusType == 5)
+                deferred.reject(new Error({status: response.status, body: response.body}));
+            else {
+                deferred.resolve(response.body);
+            }
+        });
+
+    return deferred.promise;
+};
+
 module.exports = ProjectUtils;